Switch to using gopher-lua and gopher-luar.
No longer use golua, which is a wrapper around the C Lua API. This way we don't have to deal with
any cgo issues that may crop up, and we are now pure Go.

We also move the main binary (package main) into a cmd subdirectory for cleaner packaging going
forward. The rest of the code/packages probably need to be cleaned up/rethought as well.