Slightly rewrite cli flag handling and remove chroot functionality.
Fix flag handling so that it correctly allows for overriding values from the config file.

Also remove chroot functionality. Since Go lacks the ability to correctly setuid/setgid (drop
priviledges), and we never want to run the game as root, the chroot functionality is mostly useless.
The correct way of handling security from that perspective is with a container or jail. Most users
of this code will just be running the binary locally out of the source directory though, so even
then chroot was an unneeded feature.

We can investigate technologies such as docker/jails/etc and capabilities for more secure
deployments in the future. If we can use linux capabilities, we can probably reintroduce the chroot
functionality since it no longer requires root priviledges.