Split out client networking functionality into a separate proxy process.
We're going to try something different with this version of Ataxia.
Using a separate proxy process for client networking, we can do away
with the hotboot/copyover functionality in the main engine process. This
way we can reboot the main engine without having to keep track of client
connection state. The proxy process with manage all client sockets and
login functionality, and will communicate to the main engine via some
custom protocol (protobuf? gRPC?).