Large change in this commit. Convert ataxia-proxy to Go.
All of the Go code is copied from the previous version of Ataxia, though
pruned of the game code (including all Lua code). This leaves most of
the network code, which still works to accept connections up to a
certain point.

This commit also adds all the tooling support for Go from the previous
version of ataxia. Including adding such support to the Makefile.

There's also some miscellaneous such as updating the README, updating
the Rust dependencies, etc.