Clean up the Makefile so it does less work.
Now that we don't have to compile Go and Rust separately, just rely on
cargo build to compile both binaries automatically. This also prevents
cargo from needing to compile the code twice. We can also combine
both lint targets, since due to the way we've designed the Rust code
(tiny source files for the "binaries", fat library), it makes no sense
to run the lints twice, once for each binary.

I left the two targets for proxy and engine in place, in case we want to
compile just one of the binaries without overwriting the other one. This
may change in the future, perhaps we can add an install target that
copies the binaries to bin/.