Add more clippy checks (pendantic), and fix up the errors that resulted.
I'm restructuring things a bit, now that I've moved the proxy back to
Rust. It required a bit of rearranging of the config module and config
files. Now we have two config files instead of one combined file. This
is just easier since the two daemons have different PID and log files,
even though they share other configuration items.

In the future we may collapse this back into a single shared config
file, and specify the PID and log files on the command line, within the
startup/supervisor script. We'll have to do that eventually anyway to
allow for multiple instances/versions of the daemons to run (for
example, for testing new code on the same server).